Move ".server" from file to folder level
