+++ app/routes/$show.[styles.css].ts
@@ -1,6 +1,6 @@
 import type { LoaderFunction } from '@remix-run/node';
 
-import { db } from '~/db/db.server';
+import { db } from '~/db.server/db';
 
 const notFound = () => new Response('Not Found', { status: 404 });
