+++ app/routes/admin.shows.new.tsx
@@ -8,7 +8,7 @@ import type { FC } from 'react';
 import { validationError } from 'remix-validated-form';
 
 import { redirectToLogin } from '~/auth/redirect-to-login.server';
-import { db } from '~/db/db.server';
+import { db } from '~/db.server/db';
 import { NewShowForm } from '~/forms/show/forms';
 import { makeServerValidator } from '~/forms/show/validator.server';
